+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.1.2 (2017-08-16)
+
+- Add Document.format() to get the document's formatted string.
+- Properly parse double quotes within a cell.
+- Add error messages for unfinished quotes within a cell.
+
 ### 1.1.1 (2017-08-15)
 
 - Add functions for getting the column index and row index from a character index.